some for line

+++ src/main/java/cn/moo/trainingcollege/service/StatService.java
@@ -54,14 +54,14 @@ public interface StatService {
      * @param organId
      * @return 收入List，过去十二个月的收入，顺序为从前到后
      */
-    public List<Integer> getOrganIncomeLine(String organId);
+    public List<Double> getOrganIncomeLine(String organId);
 
     /**
      * 获得机构的参加课程人数曲线（各个月之间不用去重）
      * @param organId
      * @return 人数List，过去十二个月加入的人数，顺序为从前到后
      */
-    public List<Integer> getOrganMemberLine(String organId);
+    public List<Double> getOrganMemberLine(String organId);
 
     /**
      * 获得机构参与人数最高的课程名及其人数
@@ -98,6 +98,6 @@ public interface StatService {
      * 获得过去十二个月的加入人数
      * @return 人数List，过去十二个月加入的人数，排序为从前到后
      */
-    public List<Integer> getSiteMemberLine();
+    public List<Double> getSiteMemberLine();
 
 }
